color scheme directive
