+++ projects/countdown/src/app/services/LocalStorageService.ts
@@ -1,11 +1,9 @@
 import { Injectable } from '@angular/core';
 import { LocalStorage } from '@cocco3/utils';
-
-export const UserSettingThemes = ['auto', 'light', 'dark'] as const;
-export type UserSettingTheme = (typeof UserSettingThemes)[number];
-
-export const UserSettingFormats = ['detailed', 'days'] as const;
-export type UserSettingFormat = (typeof UserSettingFormats)[number];
+import {
+  type UserSettingFormat,
+  type UserSettingTheme,
+} from './UserSettingsService';
 
 export type AppLocalStorage = {
   settings: {
